fixed bracket creation since router caused bug. Also cleaned up some unused vars in several files. Also create PageNotFound route.
