+++ src/components/DisplayBrackets.js
@@ -1,23 +1,7 @@
-import { useEffect, useState } from 'react';
-import axios from 'axios';
 import DisplayRound from './DisplayRound'
 import CreateFirstRoundBracket from './CreateFirstRoundBracket'
 
-const DisplayBrackets = ({ tourney, tournamentId, userList, currUser }) => {
-  const [brackets, setBrackets] = useState([]);
-
-  const getBrackets = () => {
-    axios.get('https://bracketsrus.herokuapp.com/api/brackets')
-    .then((response) => {
-      const allBrackets = response.data;
-      const filteredBrackets = allBrackets.filter(bracket => bracket.tournament === tournamentId);
-      setBrackets(filteredBrackets);
-    });
-  };
-
-  useEffect(() => {
-    getBrackets();
-  }, []);
+const DisplayBrackets = ({ tourney, userList, currUser, brackets, getBrackets }) => {
 
   return (
       <>
@@ -26,13 +10,13 @@ const DisplayBrackets = ({ tourney, tournamentId, userList, currUser }) => {
           <h3>Brackets</h3>
           {brackets.map((bracket, index) => {
             return (
-              <>
+              <div key={index}>
               {index === brackets.length - 1 ?
                 <DisplayRound tourney={tourney} currUser={currUser} bracket={bracket} userList={userList} key={bracket.id} getBrackets={getBrackets} isLatestRound={true}/>
                 :
                 <DisplayRound tourney={tourney} currUser={currUser} bracket={bracket} userList={userList} key={bracket.id} getBrackets={getBrackets} isLatestRound={false}/>
               }
-              </>
+              </div>
             )
           })}
           </>
