+++ src/components/Owned.js
@@ -1,4 +1,4 @@
-import {useState, useEffect} from 'react';
+import {useEffect} from 'react';
 import axios from 'axios';
 import Edit from '../components/Edit.js';
 import {Link} from 'react-router-dom';
